web3js: Add example/accounts
