+++ web3js-ext/example/README.md
@@ -1,5 +1,6 @@
 # ethers-ext examples
 
+- [accounts](./accounts) `web3.eth.accounts` APIs
 - [accountKey](./accountKey) Klaytn AccountKey types
 - [browser-html](./browser-html) Browser extension wallets (e.g. Kaikas) interaction in plain HTML
 - [transactions](./transactions) Klaytn transaction types
